--- java/src/main/java/codingdojo/Address.java
@@ -1,37 +1,0 @@
-package codingdojo;
-
-public class Address {
-    private String street;
-    private String city;
-    private String postalCode;
-
-    public Address(String street, String city, String postalCode) {
-        this.street = street;
-        this.city = city;
-        this.postalCode = postalCode;
-    }
-
-    public String getStreet() {
-        return street;
-    }
-
-    public String getPostalCode() {
-        return postalCode;
-    }
-
-    public void setPostalCode(String postalCode) {
-        this.postalCode = postalCode;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public void setStreet(String street) {
-        this.street = street;
-    }
-}
