+++ java/src/main/java/codingdojo/datawriter/DataWriter.java
@@ -1,6 +1,6 @@
 package codingdojo.datawriter;
 
-import codingdojo.Customer;
+import codingdojo.domain.Customer;
 import java.util.Collection;
 
 public interface DataWriter {
