+++ java/src/main/java/codingdojo/domain/CustomerType.java
@@ -1,4 +1,4 @@
-package codingdojo;
+package codingdojo.domain;
 
 public enum CustomerType {
     PERSON, COMPANY
