+++ java/src/main/java/codingdojo/domain/ShoppingList.java
@@ -1,4 +1,4 @@
-package codingdojo;
+package codingdojo.domain;
 
 import java.util.Arrays;
 import java.util.List;
